Split the generalised functions into a tools script, the remaining data classes / high level models now sit in main, with general functions / tools calling on those data classes sitting in tools.
